+++ folio-service-tools-dev/src/main/java/org/folio/util/TokenUtils.java
@@ -4,8 +4,6 @@ import static org.apache.commons.lang3.ObjectUtils.defaultIfNull;
 
 import static org.folio.util.FutureUtils.failedFuture;
 
-import java.nio.charset.StandardCharsets;
-import java.util.Base64;
 import java.util.Collections;
 import java.util.Map;
 import java.util.Optional;
@@ -13,18 +11,14 @@ import java.util.concurrent.CompletableFuture;
 
 import javax.ws.rs.NotAuthorizedException;
 
-import io.vertx.core.json.JsonObject;
 import org.apache.commons.collections4.map.CaseInsensitiveMap;
 import org.apache.commons.lang3.StringUtils;
-
+import org.folio.okapi.common.OkapiToken;
 import org.folio.okapi.common.XOkapiHeaders;
 
 public final class TokenUtils {
 
   private static final String INVALID_TOKEN_MESSAGE = "Invalid token";
-
-  private static final String USER_ID_KEY = "user_id";
-  private static final String USERNAME_KEY = "sub";
 
   private TokenUtils() {
   }
@@ -37,9 +31,13 @@ public final class TokenUtils {
    */
   public static Optional<UserInfo> userInfoFromToken(String token) {
     try {
-      String[] split = token.split("\\.");
-      JsonObject j = new JsonObject(getJson(split[1]));
-      return Optional.of(new UserInfo(j.getString(USER_ID_KEY), j.getString(USERNAME_KEY)));
+      var okapiToken = new OkapiToken(token);
+      if (okapiToken.getPayloadWithoutValidation() == null) {
+        return Optional.empty();
+      }
+      return Optional.of(new UserInfo(
+          okapiToken.getUserIdWithoutValidation(),
+          okapiToken.getUsernameWithoutValidation()));
     } catch (Exception e) {
       return Optional.empty();
     }
@@ -55,10 +53,5 @@ public final class TokenUtils {
     Map<String, String> h = new CaseInsensitiveMap<>(defaultIfNull(okapiHeaders, Collections.emptyMap()));
 
     return fetchUserInfo(h.get(XOkapiHeaders.TOKEN));
-  }
-
-  private static String getJson(String strEncoded) {
-    byte[] decodedBytes = Base64.getDecoder().decode(strEncoded);
-    return new String(decodedBytes, StandardCharsets.UTF_8);
   }
 }