+++ NEWS.md
@@ -11,6 +11,7 @@
 
 ### Bug fixes
 * Do not delete kafka topics if collection topic is enabled ([FST-77](https://folio-org.atlassian.net/browse/FST-77))
+* Fix incorrect base64 decoding in TokenUtils ([FST-80](https://folio-org.atlassian.net/browse/FST-80))
 
 ### Tech Dept
 * Description ([ISSUE](https://folio-org.atlassian.net/browse/ISSUE))